+++ packages/node/src/indexer/types.ts
@@ -5,6 +5,8 @@ import { ApiPromise } from '@polkadot/api';
 import { ApiDecoration } from '@polkadot/api/types';
 import type { HexString } from '@polkadot/util/types';
 import {
+  BlockHeader,
+  LightSubstrateEvent,
   SubstrateBlock,
   SubstrateEvent,
   SubstrateExtrinsic,
@@ -16,6 +18,17 @@ export interface BlockContent {
   events: SubstrateEvent[];
 }
 
+export interface LightBlockContent {
+  block: BlockHeader; // A subset of SubstrateBlock
+  events: LightSubstrateEvent[];
+}
+
 export type BestBlocks = Record<number, HexString>;
 
 export type ApiAt = ApiDecoration<'promise'> & { rpc: ApiPromise['rpc'] };
+
+export function isFullBlock(
+  block: BlockContent | LightBlockContent,
+): block is BlockContent {
+  return (block as BlockContent).extrinsics !== undefined;
+}